+++ CHANGELOG.md
@@ -29,9 +29,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - MDR URN in attribute template
 - Child fhir path
 - Spring Boot 2.7.7
+- Anonym Attribute Template
 - Converter Session
 - Apache POI
 - Excel Converter
+- Anonym Attributes
+- Operation in Attribute
+- EXTRACT_RELATIVE_ID attribute operation
 
 ### Changed
 - Rename Conversion Template as Converter Template